+++ src/Entity/Player.php
@@ -21,7 +21,7 @@ class Player
     private $name;
 
     /**
-     * @var int
+     * @var float
      */
     private $chips;
 
@@ -53,36 +53,36 @@ class Player
     }
 
     /**
-     * @return int
+     * @return float
      */
-    public function getChips(): int
+    public function getChips(): float
     {
         return $this->chips;
     }
 
     /**
-     * @param int $chips
+     * @param float $chips
      */
-    public function setChips(int $chips): void
+    public function setChips(float $chips): void
     {
         $this->chips = $chips;
     }
 
 
     /**
-     * @param int $amount
-     * @return int
+     * @param float $amount
+     * @return float
      */
-    public function betChips(int $amount): int
+    public function betChips(float $amount): float
     {
         $this->chips -= $amount;
         return $amount;
     }
 
     /**
-     * @param int $amount
+     * @param float $amount
      */
-    public function winChips(int $amount): void
+    public function winChips(float $amount): void
     {
         $this->chips += $amount;
     }
